fix(merge_concat_tests): merge in one file and cleaning
